Merge facebook account with existing local account
After a successful login using Facebook, we will have the person's email
address, Facebook ID, and access token. we search for an existing
account that has been created with that same email address. If one
exists, we merge the two accounts and add the Facebook info to the
existing account. If there is not an existing user in our database with
that email address, then we create a new user.